module: check .node_preloadrc for preload modules
Adds a new feature that checks .node_preloadrc for a list of modules
to preload.  One module per line and each one can be the absolute
path or just the name of the module.

Two locations will be check for the rc file: HOMEDIR or CWD.
The one in HOMEDIR must specify modules under the global node_modules
dir only, and the one in CWD must specify modules under the dir
CWD/node_modules only.

This is smilar to what can be specified by the -r command line
option except the modules are limited to the global or local
node_modules only.

This feature must be opted-in with setting the env var
NODE_PRELOADRC_ENABLE to a non-falsy value.

Implements: https://github.com/nodejs/node/issues/11853